support stop_grace_period in launched container
I have some containers that execute long-running jobs, and I would like them to be able to finish, rather than have docker kill the container after the default 10 seconds.  E.g., if a job typically takes 10 minutes to run, and a job has been running for 9m40s when the service is stopped or updated, I want to let it finish.

Note that when the new the LAUNCH_STOP_GRACE_PERIOD option is specified, "stop_grace_period: <duration>" also needs to be added to the stack.yml file, assuming you want the swarm-launcher container to stick around until the launched container stops.  And the duration specified in the yaml file should be slightly longer than the LAUNCH_STOP_GRACE_PERIOD value.